requirements.txt
requests
beautifulsoup4
tkinter
Pillow
string
yfinance
selenium
